+++ src/main/java/com/verdantartifice/primalmagic/client/events/ClientModLifecycleEvents.java
@@ -6,6 +6,7 @@ import com.verdantartifice.primalmagic.client.gui.AnalysisTableScreen;
 import com.verdantartifice.primalmagic.client.gui.ArcaneWorkbenchScreen;
 import com.verdantartifice.primalmagic.client.gui.CalcinatorScreen;
 import com.verdantartifice.primalmagic.client.gui.ConcocterScreen;
+import com.verdantartifice.primalmagic.client.gui.EssenceTransmuterScreen;
 import com.verdantartifice.primalmagic.client.gui.GrimoireScreen;
 import com.verdantartifice.primalmagic.client.gui.HoneyExtractorScreen;
 import com.verdantartifice.primalmagic.client.gui.ResearchTableScreen;
@@ -92,6 +93,7 @@ public class ClientModLifecycleEvents {
         MenuScreens.register(ContainersPM.RUNECARVING_TABLE.get(), RunecarvingTableScreen::new);
         MenuScreens.register(ContainersPM.HONEY_EXTRACTOR.get(), HoneyExtractorScreen::new);
         MenuScreens.register(ContainersPM.CONCOCTER.get(), ConcocterScreen::new);
+        MenuScreens.register(ContainersPM.ESSENCE_TRANSMUTER.get(), EssenceTransmuterScreen::new);
     }
     
     private static void registerTERs() {